refactor(web): extract ShortcutHelpDialog from App.tsx
Mechanical move of SHORTCUT_GROUPS and the help modal JSX into
components/app/ShortcutHelpDialog.tsx (props: open/onClose). The known
g/G duplication in the shortcut data is left for CS-21.

Issue: CS-18 (step 3/3)
